Добавляет компонент MovieCardFull
